refactor import progress tracker
