+++ api/src/modules/events/app-events.module.ts
@@ -1,10 +1,10 @@
 import { Global, Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
-import { ImportProgressHandler } from 'modules/events/import-data/import-progress.handler';
-import { ImportProgressEmitter } from 'modules/events/import-data/import-progress.emitter';
+import { ImportProgressHandler } from 'modules/events/import-data-progress/import-progress.handler';
+import { ImportProgressEmitter } from 'modules/events/import-data-progress/import-progress.emitter';
 import { WebSocketsModule } from 'modules/notifications/websockets/websockets.module';
-import { ImportProgressSocket } from 'modules/events/import-data/import-progress.socket';
-import { ImportProgressTrackerFactory } from 'modules/events/import-data/import-progress.tracker.factory';
+import { ImportProgressSocket } from 'modules/events/import-data-progress/import-progress.socket';
+import { ImportProgressTrackerFactory } from 'modules/events/import-data-progress/import-progress.tracker.factory';
 
 @Global()
 @Module({
