+++ api/src/modules/geo-coding/geo-coding.service.ts
@@ -11,7 +11,7 @@ import {
 import { GeoCodingAbstractClass } from 'modules/geo-coding/geo-coding-abstract-class';
 import { AdminRegionOfProductionService } from 'modules/geo-coding/strategies/admin-region-of-production.service';
 import { GeoCodingProgressTracker } from 'modules/geo-coding/progress-tracker/geo-coding.progress-tracker';
-import { ImportProgressTrackerFactory } from '../events/import-data/import-progress.tracker.factory';
+import { ImportProgressTrackerFactory } from '../events/import-data-progress/import-progress.tracker.factory';
 
 interface locationInfo {
   locationAddressInput?: string;
