+++ api/src/modules/geo-coding/progress-tracker/geo-coding.progress-tracker.ts
@@ -1,4 +1,4 @@
-import { ImportProgressEmitter } from 'modules/events/import-data/import-progress.emitter';
+import { ImportProgressEmitter } from 'modules/events/import-data-progress/import-progress.emitter';
 
 export class GeoCodingProgressTracker {
   totalLocations: number;
