+++ api/src/modules/sourcing-locations/progress-tracker/sourcing-data.progress-tracker.ts
@@ -1,4 +1,4 @@
-import { ImportProgressEmitter } from 'modules/events/import-data/import-progress.emitter';
+import { ImportProgressEmitter } from 'modules/events/import-data-progress/import-progress.emitter';
 
 export class SourcingDataImportProgressTracker {
   totalRecords: number;
